pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_indexer"
version = "0.0.1"
authors = [
  { name="flowpoint" },
]
description = "a small ai indexing and search tool"
readme = "readme.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
license = {text = "MIT License"}
dependencies = [
    'click',
    'lmdb',
    'tokenizers',
    'onnxruntime',
    'faiss-cpu',
    'zarr',
    'more_itertools',
]

[project.optional-dependencies]
dev = [
    'ipython',
    'python-lsp-server',
    'mypy',
    'pytest',
]

quality_bench = [
    'mteb'
]

flexible = [
'torch',
'transformers'
]

[project.scripts]
ai_indexer = "ai_indexer.cli:cli"

#[project.urls]
#"Homepage" = "https://github.com/pypa/sampleproject"
#"Bug Tracker" = "https://github.com/pypa/sampleproject/issues"


